WIP (fill in TBD) Add a workflow to initialize the Git notes
So far, GitGitGadget has supported one project, and one single project
only: the Git project.

But the Git project is far from the only one that prides itself by doing
patch review on a mailing list.

To allow GitGitGadget's convenience of PRs being sent to mailing lists
to benefit other projects, too, here is a workflow that will have to be
triggered to initialize the prerequisite Git notes in the `pr-repo`.

For details about setting up GitGitGadget to support a project other
than Git, see TBD.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>