WIP (fill in TBD) Use `vars.CONFIG` with all of GitGitGadget's GitHub Actions
The GitHub Actions of GitGitGadget just learned the trick to accept the
project configuration as an input parameter. The idea is to store this
configuration in the repository variable called `CONFIG`. Here is the
code change to make use of this repository variable.

Obviously, this means that the workflows should no longer run in forks
without said configuration.

For details about configuring GitGitGadget to support projects other
than Git, see TBD.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>