Use the `{upstream,pr,test}-repo` information from the repository config
We just started to pass the project-specific configuration stored in the
`CONFIG` repository variable to the GitHub Actions of GitGitGadget.

That configuration contains information about the correct repositories
to target (which might not be `gitgitgadget/git` and friends). Let's use
this information when obtaining (and using) the installation access
tokens.

This is especially necessary when creating and updating that Check Run
in `handle-pr-push` and in `handle-pr-comment`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>