config: add newly-required attributes
As of 80fb71e1 (Extend the `IConfig` interface as needed for 3rd-party
projects, 2025-08-31), there are a couple more non-optional attributes
in the project-specific settings.

Let's adjust them here, in the `config` branch, which serves as the
source of truth for the project-specific settings used by this
particular fork of `gitgitgadget-workflows`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>