config: adjust to the new schema
In https://github.com/gitgitgadget/gitgitgadget/pull/1991, I adjusted
and augmented the `IConfig` interface a bit. Concretely, I renamed
`repo.baseOwner` to `repo.upstreamOwner` in c374c542 (IConfig: rename
the attribute defining the `upstream-repo`'s org, 2025-09-08) and moved
`repo.owners` to `app.installedOn` in 907807fb (IConfig: move
`repo.owners` to a better place, 2025-09-08).

Let's adjust to that new schema.

Note: the `baseOwner` and `owners` attributes are left as-are, for the
transitional period until that PR and its friends are merged.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>