Use a less git.git-centric name for the read-only GitHub App information
When working on `git/git`, GitGitGadget needs to use a GitHub App that
lacks write permissions, i.e. a read-only App (as opposed to the primary
GitHub App that does have write permissions but is installed on
`gitgitgadget/git`, not on `git/git`).

The information to act as that App is stored in a pair of repository
secrets: `GITGITGADGET_GIT_GITHUB_APP_ID` and
`GITGITGADGET_GIT_GITHUB_APP_PRIVATE_KEY`. The `_GIT_` infix, however,
strongly suggests the repository _name_ rather than the _role_, and
reflects GitGitGadget's primary scenario where only the Git project is
supported.

In preparation for supporting other projects, too, let's use the infix
`_READONLY_` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>