Add `project` to the config
As of 875d97d1 (Set `project` in the `defaultConfig`, 2025-09-08), even
the `project`-specific part of the config (defining how to send the
email to the Git project) is no longer hard-coded in the Typescript code
and must be specified via the regular `IConfig`. So let's add that part
here, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>